- fix broken unit tests due to EventFactory after moving Event model to artwork modules
